### 2016-11-13 #110 @whedon
[PRE REVIEW]: finreportr - Financial Data from U.S. Securities and Exchange Commission
**Submitting author:** @sewardlee337 (<a href="http://orcid.org/0000-0001-5583-035X">Seward Lee</a>)
**Repository:** <a href="https://github.com/sewardlee337/finreportr" target ="_blank">https://github.com/sewardlee337/finreportr</a>
**Version:** v1.0.1

Thanks for submitting your paper to JOSS sewardlee337. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
