### 2016-11-13 #110 @arfon
👋 @sewardlee337, thanks for submitting this package to JOSS. Before we get started finding an editor and reviewer could you confirm that this software has a research application? In particular, our submission guidelines state:

> - Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

What is the research challenge that this software is designed to address?
