### 2016-11-13 #110 @sewardlee337
@arfon thank you for prepping my package for review!

My package contributes to research in empirical finance by automating the download of data relating to financial health and transactions of publicly-listed companies, thus speeding-up and simplifying the financial research workflow.

My paper summary has been updated accordingly.
