### 2016-11-14 #110 @arfon
:wave: @jsta - do you think you'd be able to review this submission? It's in R and looks to be some kind of API client library.
