### 2016-11-14 0aaacd2
Edit JOSS paper summary
