### 2016-11-14 #119 @whedon
[REVIEW]: finreportr - Financial Data from U.S. Securities and Exchange Commission
**Submitting author:** @sewardlee337 (<a href="http://orcid.org/0000-0001-5583-035X">Seward Lee</a>)
**Repository:** <a href="https://github.com/sewardlee337/finreportr" target ="_blank">https://github.com/sewardlee337/finreportr</a>
**Version:** v1.0.1

- [x] **Authorship:** Has the submitting author (sewardlee337) made major contributions to the software?
