### 2016-11-14 #119 @arfon
@jsta - many thanks for agreeing to review this submission. Please take a look at the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) and check off items on the checklist above as you progress.
