### 2016-11-14 sewardlee337/finreportr#1 @jsta
No filings available for given year
My run of the example README.md code for `GetIncome`, `GetBalanceSheet`, and `GetCashFlow`
does not produce the expected output. For example, running `GetIncome("GOOG", 2015)` produces:

    Error in GetAccessionNo(symbol, year, foreign = FALSE) :
    no filings available for given year

https://github.com/openjournals/joss-reviews/issues/119
