+++ README.md
@@ -137,31 +137,31 @@ where
 #### Example:
 
 ```
-###   Fetch first 20 lines of Google's income statement in the Form 10-K published in 2015.
+###   Fetch first 20 lines of Google's income statement in the Form 10-K published in 2016.
 
-> head(GetIncome("GOOG", 2015), 20)
+> head(GetIncome("GOOG", 2016), 20)
 
                                Metric Units      Amount  startDate    endDate
-1                            Revenues   usd 46039000000 2012-01-01 2012-12-31
-2                            Revenues   usd 55519000000 2013-01-01 2013-12-31
-3                            Revenues   usd 66001000000 2014-01-01 2014-12-31
-4                     Cost of Revenue   usd 17176000000 2012-01-01 2012-12-31
-5                     Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
-6                     Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
-7    Research and Development Expense   usd  6083000000 2012-01-01 2012-12-31
-8    Research and Development Expense   usd  7137000000 2013-01-01 2013-12-31
-9    Research and Development Expense   usd  9832000000 2014-01-01 2014-12-31
-10      Selling and Marketing Expense   usd  5465000000 2012-01-01 2012-12-31
-11      Selling and Marketing Expense   usd  6554000000 2013-01-01 2013-12-31
-12      Selling and Marketing Expense   usd  8131000000 2014-01-01 2014-12-31
-13 General and Administrative Expense   usd  3481000000 2012-01-01 2012-12-31
-14 General and Administrative Expense   usd  4432000000 2013-01-01 2013-12-31
-15 General and Administrative Expense   usd  5851000000 2014-01-01 2014-12-31
-16                 Costs and Expenses   usd 32205000000 2012-01-01 2012-12-31
-17                 Costs and Expenses   usd 40116000000 2013-01-01 2013-12-31
-18                 Costs and Expenses   usd 49505000000 2014-01-01 2014-12-31
-19            Operating Income (Loss)   usd 13834000000 2012-01-01 2012-12-31
-20            Operating Income (Loss)   usd 15403000000 2013-01-01 2013-12-31
+1                            Revenues   usd 55519000000 2013-01-01 2013-12-31
+2                            Revenues   usd 66001000000 2014-01-01 2014-12-31
+3                            Revenues   usd 74989000000 2015-01-01 2015-12-31
+4                     Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
+5                     Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
+6                     Cost of Revenue   usd 28164000000 2015-01-01 2015-12-31
+7    Research and Development Expense   usd  7137000000 2013-01-01 2013-12-31
+8    Research and Development Expense   usd  9832000000 2014-01-01 2014-12-31
+9    Research and Development Expense   usd 12282000000 2015-01-01 2015-12-31
+10      Selling and Marketing Expense   usd  6554000000 2013-01-01 2013-12-31
+11      Selling and Marketing Expense   usd  8131000000 2014-01-01 2014-12-31
+12      Selling and Marketing Expense   usd  9047000000 2015-01-01 2015-12-31
+13 General and Administrative Expense   usd  4432000000 2013-01-01 2013-12-31
+14 General and Administrative Expense   usd  5851000000 2014-01-01 2014-12-31
+15 General and Administrative Expense   usd  6136000000 2015-01-01 2015-12-31
+16                 Costs and Expenses   usd 40116000000 2013-01-01 2013-12-31
+17                 Costs and Expenses   usd 49505000000 2014-01-01 2014-12-31
[…]
