### 2016-11-14 sewardlee337/finreportr#1 @sewardlee337
Thank you for the catch. The example in the README is outdated because of GOOG's change in corporate structure in 2015 -- the SEC website no longer hosts GOOG's 2015 annual report.

I have changed examples in README.md to 2016 accordingly (e.g. `GetIncome("GOOG", 2016)`).
