### 2016-11-15 sewardlee337/finreportr#1 @jsta
`GetIncome("GOOG", 2016)` appears to download files to a `XBRLcache` folder but returns the following error:

> Error in fileFromCache(file) :
>   Error in download.file(file, cached.file, method = "auto", quiet = !verbose) :
>   cannot download all files
>
> In addition: Warning message:
> In download.file(file, cached.file, method = "auto", quiet = !verbose) :
>   URL 'https://www.sec.gov/Archives/edgar/data/1652044/000165204416000012/goog-20151231_pre.xml': status was '404 Not Found'
