### 2016-11-15 sewardlee337/finreportr#1 @jsta
`sessionInfo()`

``` r
R version 3.3.2 (2016-10-31)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.1 LTS

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] finreportr_1.0.1     devtools_1.12.0.9000

 [1] Rcpp_0.12.7        XML_3.98-1.5       digest_0.6.10      dplyr_0.5.0.9000
 [5] withr_1.0.2        assertthat_0.1     XBRL_0.99.17       R6_2.2.0
 [9] DBI_0.5-1          magrittr_1.5       httr_1.2.1         stringi_1.1.2
[13] lazyeval_0.2.0     curl_2.2           xml2_1.0.0.9001    tools_3.3.2
[17] stringr_1.1.0      selectr_0.3-0      pkgload_0.0.0.9000 rvest_0.3.2
[21] memoise_1.0.0      tibble_1.2
```
