### 2016-11-15 sewardlee337/finreportr#1 @sewardlee337
Thanks! I will look into it and try to reproduce this error to figure out what is going on.
