### 2016-11-16 #119 @arfon
@sewardlee337 - could you please move all of the references into the `paper.bib` file so that we can compile the paper?
