### 2016-11-16 a01c577
Add paper.bib for JOSS
