+++ paper/paper.bib
@@ -1,0 +1,25 @@
+@Manual{R-base,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2016},
+  url = {https://www.R-project.org/},
+}
+
+@Manual{R-XBRL,
+  title = {XBRL: Extraction of Business Financial Information from 'XBRL' Documents},
+  author = {Roberto Bertolusso and Marek Kimmel},
+  year = {2016},
+  note = {R package version 0.99.17},
+  url = {https://CRAN.R-project.org/package=XBRL},
+}
+
+@Manual{R-quantmod,
+  title = {quantmod: Quantitative Finacial Modelling Framework},
+  author = {Jeffrey A. Ryan and Joshua M. Ulrich and Wouter Thielen},
+  year = {2016},
+  note = {R package version 0.4-7},
+  url = {https://CRAN.R-project.org/package=quantmod},
+}
+
