### 2016-11-16 ca62c5a
Edit paper.md citations
