+++ paper/paper.md
@@ -12,8 +12,8 @@ bibliography: paper.bib
 
 # Summary
 
-'finreportr' is an R package (R Core Team 2016) that contributes to empirical finance research by automating the download of data relating to the health and transactions of publicly-listed companies, thus simplifying the workflow of financial analysts. It downloads and processes XBRL documents containing financial data from the U.S. Securities and Exchange Commission's EDGAR database. Using a suite of simple functions, data from the Statement of Operations, Balance Sheet, and Statement of Cash Flow of Form 10-Ks may be downloaded, processed, and returned to the user in R dataframes.
+'finreportr' is an R package [@R-base] that contributes to empirical finance research by automating the download of data relating to the health and transactions of publicly-listed companies, thus simplifying the workflow of financial analysts. It downloads and processes XBRL documents containing financial data from the U.S. Securities and Exchange Commission's EDGAR database. Using a suite of simple functions, data from the Statement of Operations, Balance Sheet, and Statement of Cash Flow of Form 10-Ks may be downloaded, processed, and returned to the user in R dataframes.
 
-The design of functions included in 'finreportr' is inspired by the quantitative finance package 'quantmod' (Ryan, Ulrich, and Thielsen 2016). XBRL-processing capabilities of 'finreportr' are mainly powered by the package 'XBRL' (Bertolusso and Kimmel 2016) under the hood.
+The design of functions included in 'finreportr' is inspired by the quantitative finance package 'quantmod' [@R-quantmod]. XBRL-processing capabilities of 'finreportr' are mainly powered by the package 'XBRL' [@R-XBRL] under the hood.
 
 # References
