### 2016-11-16 #119 @sewardlee337
@arfon - references have been added to paper.bib. Thanks!
