### 2016-11-16 sewardlee337/finreportr#1 @sewardlee337
I've tried it on Windows and Ubuntu, and it seems to work on my end.

`sessionInfo`

``` r
R version 3.2.3 (2015-12-10)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.1 LTS

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8    LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] finreportr_1.0.1 devtools_1.12.0

 [1] magrittr_1.5   R6_2.2.0       assertthat_0.1 DBI_0.5-1      tools_3.2.3    withr_1.0.2    dplyr_0.5.0    tibble_1.2
 [9] curl_2.2       Rcpp_0.12.7    memoise_1.0.0  digest_0.6.10
```

@jsta - Could you try deleting the `XBRLcache` folder and try running `GetIncome("GOOG", 2016)` again? If that still does not work, could you send me a list of file contents in the folder, so that I may see what's going on?
