### 2016-11-17 sewardlee337/finreportr#1 @jsta
Still getting an error. I tracked down the source of the error to the line in the `GetInstFile` function calling the `XBRL::xbrlDoAll` function. I cannot even run the examples from the man page for `XBRL::xbrlDoAll`. It fails with the same error.
