### 2016-11-17 sewardlee337/finreportr#1 @sewardlee337
`xbrlDoAll` calls a function in the `XBRL` package.

Can you try running `XBRL::xbrlDoALL('https://www.sec.gov/Archives/edgar/data/1288776/000165204416000012/goog-20151231.xml')`?  This will help me determine if the problem is in the `XBRL` package.

(This function from the `XBRL` package will download  a lot of files, which you may want to delete afterwards.)
