### 2016-11-17 sewardlee337/finreportr#1 @jsta
Thats what I was saying in my earlier comment. I am having trouble with the `XBRL` package itself. Your example gives me the same error message as I originally reported. Too bad `XBRL` doesn't have an official Github page to report issues.
