### 2016-11-17 sewardlee337/finreportr#1 @jsta
Why is `XBRL` not in your `sessionInfo()` under `loaded via a namespace`?
