### 2016-11-17 sewardlee337/finreportr#1 @sewardlee337
After I run `GetIncome()`, `XBRL` appears under `loaded via a namespace`:

``` R
R version 3.2.3 (2015-12-10)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.1 LTS

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8    LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] finreportr_1.0.1

 [1] Rcpp_0.12.7    XML_3.98-1.5   dplyr_0.5.0    assertthat_0.1 XBRL_0.99.17   R6_2.2.0       DBI_0.5-1      magrittr_1.5
 [9] httr_1.2.1     stringi_1.1.2  curl_2.2       lazyeval_0.2.0 xml2_1.0.0     tools_3.2.3    stringr_1.1.0  selectr_0.3-0
[17] rvest_0.3.2    tibble_1.2
```

I am corresponding with the author of the `XBRL` package directly via email to get some insight into why GetIncome() doesn't run properly in your R session...
