### 2016-11-20 sewardlee337/finreportr#1 @sewardlee337
The author of `XBRL` also runs Ubuntu 16.04. His only suggestion:

> Maybe running an:
> update.packages()
> may help.

If it works on some computers and R sessions, but not others, I suspect a settings or configuration problem. Will continue to investigate.
