### 2016-11-21 #119 @arfon
@jsta - here's the compiled paper: [10.21105.joss.00119.pdf](https://github.com/openjournals/joss-reviews/files/603509/10.21105.joss.00119.pdf)
