### 2016-11-26 sewardlee337/finreportr#1 @sewardlee337
@jsta - I am still trying to diagnose why the `XBRL` package does not work on your computer/R session.

Can you type the following into your terminal, and show me the output?
```
curl -I -v https://www.sec.gov/Archives/edgar/data/1652044/000165204416000012/goog-20151231_pre.xml
```
I am hoping this will give me more information to diagnose the underlying issue.
