### 2016-11-27 sewardlee337/finreportr#1 @jsta
I think I solved the downloading issue using the fix [here](https://stackoverflow.com/questions/29752418/how-to-download-a-file-using-cookies-with-rcurl). Unfortunately, now I'm getting an error with `fixFileName` and I'm not familiar enough with Rcpp to debug whats going on in `xbrlGetSchemaName.cpp`.

Using `options(error = recover)`:
> Error in if (substr(file.name, 1, 5) != "http:") { :
  argument is of length zero

> Enter a frame number, or 0 to exit

> 1: xbrlDoAll(inst, cache.dir = "XBRLcache", prefix.out = "out", verbose = TRUE)
2: xbrlDoAll.R#30: xbrl$processSchema(xbrl$getSchemaName())
3: XBRL.R#115: cat("Schema: ", file, "\n")
4: xbrl$getSchemaName()
5: XBRL.R#110: fixFileName(dname.inst, .Call("xbrlGetSchemaName", doc.inst, PACK
