### 2016-11-27 sewardlee337/finreportr#1 @jsta
Ok, it turns out that the `download.file` issue described [here](http://stackoverflow.com/a/26878648/3362993) was the problem. Installing `XBLR` from my fork with `devtools::install_github("jsta/XBLR")` fixed the problem completely!

`head(GetIncome("GOOG", 2016))`

> Metric Units      Amount  startDate    endDate
> 1        Revenues   usd 55519000000 2013-01-01 2013-12-31
2        Revenues   usd 66001000000 2014-01-01 2014-12-31
3        Revenues   usd 74989000000 2015-01-01 2015-12-31
4 Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
5 Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
6 Cost of Revenue   usd 28164000000 2015-01-01 2015-12-31
