### 2016-11-28 sewardlee337/finreportr#1 @sewardlee337
@jsta - I've done some research on the Stack Overflow solution. It seems like the issue has to do with the method that `download.file()` uses to download the XBRL files (see [official help page](http://stat.ethz.ch/R-manual/R-devel/library/utils/html/download.file.html)). `method = "curl"` seems to be what works for you.

I've never had a problem with the default setting `method = "auto"` after testing the `XBRL` package on three computers (Windows and Ubuntu), so I suspected that it's an issue with global options settings in your R session. I've tried toggling through different `download.file()` methods in my options settings, but still couldn't reproduce your bug. For example:

```R
> ### method = "wget"
> getOption("download.file.method")
[1] "wget"
>
> head(GetIncome("GOOG", 2016))
           Metric Units      Amount  startDate    endDate
1        Revenues   usd 55519000000 2013-01-01 2013-12-31
2        Revenues   usd 66001000000 2014-01-01 2014-12-31
3        Revenues   usd 74989000000 2015-01-01 2015-12-31
4 Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
5 Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
6 Cost of Revenue   usd 28164000000 2015-01-01 2015-12-31
>
> ### method = "internal"
> options(download.file.method = "internal")
> head(GetIncome("GOOG", 2016))
           Metric Units      Amount  startDate    endDate
1        Revenues   usd 55519000000 2013-01-01 2013-12-31
2        Revenues   usd 66001000000 2014-01-01 2014-12-31
3        Revenues   usd 74989000000 2015-01-01 2015-12-31
4 Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
5 Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
6 Cost of Revenue   usd 28164000000 2015-01-01 2015-12-31
>
> ### method = "auto"
> options(download.file.method = "auto")
> head(GetIncome("GOOG", 2016))
           Metric Units      Amount  startDate    endDate
1        Revenues   usd 55519000000 2013-01-01 2013-12-31
2        Revenues   usd 66001000000 2014-01-01 2014-12-31
3        Revenues   usd 74989000000 2015-01-01 2015-12-31
4 Cost of Revenue   usd 21993000000 2013-01-01 2013-12-31
5 Cost of Revenue   usd 25691000000 2014-01-01 2014-12-31
6 Cost of Revenue   usd 28164000000 2015-01-01 2015-12-31
```
Therefore, I currently suspect that what's causing problems on your end is some network restriction that specifically you are subjected to.

If this is the case, how should we proceed?
