### 2016-11-30 sewardlee337/finreportr#1 @jsta
Hmm, without my `XBRL` fork I am getting the same error in a Docker instance of [this image](https://hub.docker.com/r/achubaty/r-spatial-base/). I am dialed into a remote server on a separate network.

> R version 3.3.2 (2016-10-31)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Debian GNU/Linux stretch/sid

> locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

> attached base packages:

> other attached packages:
[1] finreportr_1.0.1 devtools_1.12.0

> loaded via a namespace (and not attached):
 [1] Rcpp_0.12.8    XML_3.98-1.5   digest_0.6.10  dplyr_0.5.0    withr_1.0.2
 [6] assertthat_0.1 XBRL_0.99.17   R6_2.2.0       DBI_0.5-1      git2r_0.16.0
[11] magrittr_1.5   httr_1.2.1     stringi_1.1.2  lazyeval_0.2.0 curl_2.3
[16] xml2_1.0.0     tools_3.3.2    stringr_1.1.0  selectr_0.3-0  rvest_0.3.2
[21] memoise_1.0.0  knitr_1.15.1   tibble_1.2

In my opinion the ideal would be to have more tests and see if travis has the same error.

As far as the JOSS review, I understand that there is not much you can do about an error in a dependency. Now that I have it working with my `XBRL` fork, I will proceed with the review and I suppose get the opinion of the JOSS editor afterwards...
