### 2016-11-30 #119 @jsta
The `LICENSE` file is incomplete.

As of my last check there has been no Github release.

- [x] **Authorship:** Has the submitting author (sewardlee337) made major contributions to the software?

My run of the example `README.md` code for `AnnualReports` does not match the expected output. For example, when I run `AnnualReports("GOOG")` I get:

>  filing.name filing.date         accession.no
1      10-K/A  2016-03-29 0001193125-16-520367
2        10-K  2016-02-11 0001652044-16-000012

rather than:

>    filing.name filing.date         accession.no
1         10-K  2015-02-09 0001288776-15-000008
2         10-K  2014-02-12 0001288776-14-000020
3         10-K  2013-01-29 0001193125-13-028362
4       10-K/A  2012-04-23 0001193125-12-174477
5         10-K  2012-01-26 0001193125-12-025336
6         10-K  2011-02-11 0001193125-11-032930
7         10-K  2010-02-12 0001193125-10-030774
8         10-K  2009-02-13 0001193125-09-029448
9         10-K  2008-02-15 0001193125-08-032690
10        10-K  2007-03-01 0001193125-07-044494
11        10-K  2006-03-16 0001193125-06-056598
12        10-K  2005-03-30 0001193125-05-065298

If I try the roxygen example code for `GetCashFlow("FB", 2016)` I get the following error:

> Error in fileFromCache(file) :
  Error in download.file(file, cached.file, method = "auto", quiet = !verbose) :
  cannot download all files

> In addition: Warning message:
In download.file(file, cached.file, method = "auto", quiet = !verbose) :
  URL 'https://www.sec.gov/Archives/edgar/data/1326801/000132680116000043/fb-20151231.xsd': status was '404 Not Found'

This error can be fixed by modifying the dependency `XBRL` to call `download.file(method = "curl")`. See https://github.com/jsta/XBRL/commit/9c42ada9947aadfbc2f2fe64705cabab9eddca40.

The automated `testthat` tests are extremely minimal. They only test one function. However, the manual steps described in the roxygen documentation and `README.md` should be sufficient to verify package functionality.

At the very least you should fill out a BugReports section in the `DESCRIPTION`. Some package authors also choose to include a `CONTRIBUTING.md`. See https://github.com/ropensci/rnoaa/blob/master/CONTRIBUTING.md for an example.

It seems to me that not everyone necessarily has an affiliation but going by the letter-of-the-law there is no affiliation listed in `paper.md`.

### Miscellaneous

I'm not sure how you are building this package for CRAN and avoiding the warnings about non-standard top-level files like `Read-and-delete-me` and `.RData`. These should probably be removed or appended to `Rbuildignore`.

When I run `goodpractice::gp()` on the package using [goodpractice](https://github.com/MangoTheCat/goodpractice) I get the following notes:

> ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid
    quite often. A build date will be added to the package when you
    perform `R CMD build` on it.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug
    tracker. Many online code hosting services provide bug trackers for
    free, https://github.com, https://gitlab.com, etc.
  ✖ avoid long code lines, it is bad for readability. Also, many people
    prefer editor windows that are about 80 characters wide. Try make
    your lines shorter than 80 characters

    R/annual_reports.R:18:1
    R/annual_reports.R:20:1
    R/annual_reports.R:53:1
    R/backend_utilities.R:19:1
    R/backend_utilities.R:42:1

>  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
    1:NCOL(...) expressions. They are error prone and result 1:0 if the
    expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/get_financial.R:68:17

> ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific
    functions you need.
