### 2016-11-30 #119 @jsta
@arfon I ran into somewhat of a show-stopper bug with one of `finreportr`'s dependencies (see my comment above about `XBRL`). It seems like a bug in a dependency is out of the present author's control. I was able to patch it myself to test the `finreportr` functions but the everyday user may run into frustration. How should this be handled for the review?
