### 2016-11-30 #119 @sewardlee337
@jsta - Thank you for the advice! I will start implementing some of these fixes.
