### 2016-11-30 #119 @sewardlee337
Latest updates to `finreportr` per @jsta's suggestions and JOSS review checklist:

* `BugReports` added to `DESCRIPTION` and Feedback section added to `README.md`
* MIT License full text added to `LICENSE` per JOSS requirement
* `finreportr` vers 1.0.1 released on GitHub
* Long lines of code cut down where contextually appropriate
