+++ R/get_balance_sheet.R
@@ -14,7 +14,10 @@
 
 GetBalanceSheet <- function(symbol, year) {
      
-     balance.sheet.descriptions <- c("CONSOLIDATED BALANCE SHEET", "CONSOLIDATED BALANCE SHEETS", "CONSOLIDATED STATEMENT OF FINANCIAL POSITION", "CONSOLIDATED STATEMENTS OF FINANCIAL POSITION")
+     balance.sheet.descriptions <- c("CONSOLIDATED BALANCE SHEET", 
+                                     "CONSOLIDATED BALANCE SHEETS", 
+                                     "CONSOLIDATED STATEMENT OF FINANCIAL POSITION", 
+                                     "CONSOLIDATED STATEMENTS OF FINANCIAL POSITION")
      
      GetFinancial(balance.sheet.descriptions, symbol, year)
 }
