### 2016-12-02 #119 @jsta
@arfon I think this submission is good to go as far as acceptance pending your opinion on the `XBRL` (dependency of the present submission) bug described at: https://github.com/sewardlee337/finreportr/issues/1
