### 2016-12-03 ee8a8c9
Add DOI badge from Zenodo
