+++ paper/paper.md
@@ -6,7 +6,7 @@ tags:
 authors:
  - name: Seward Lee
    orcid: 0000-0001-5583-035X
-date: 31 October 2016
+date: 3 December 2016
 bibliography: paper.bib
 ---
 
