### 2016-12-05 #119 @arfon
@jsta many thanks for the review here.

@sewardlee337 - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00119 :zap: 🚀 💥
